backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.44.1
cryptography==45.0.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
gpt4all==2.8.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
mailjet-rest==1.4.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.17.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-http-client==3.3.7
regex==2026.1.15
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tiktoken==0.12.0
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
zipp==3.23.0
